Modify UnresolvedMethods tests
